refactor: @Slf4j를 Log4j2로 변경
