+++ src/main/java/co/kr/allpick/domain/admin/dto/AdminCreateRequestDto.java
@@ -1,7 +1,6 @@
 package co.kr.allpick.domain.admin.dto;
 
 import co.kr.allpick.domain.admin.entity.Admin;
-import co.kr.allpick.domain.admin.service.AdminServiceImpl;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
